pyproject.toml
[project]
name = "smart-recipe-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
    "python-docx>=1.2.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-cors-headers>=4.3.0",
    "django-filter>=23.5",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "requests>=2.32.5",
]
